pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bear-alarm"
version = "0.2.0"
description = "Dexcom glucose monitoring with audio alerts"
requires-python = ">=3.10"
dependencies = [
    "pydexcom>=0.2.3",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pygame>=2.5.0",
    "flet[all]>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
bear-alarm = "src.main:main"
bear-alarm-cli = "src.cli:main"

[project.gui-scripts]
bear-alarm-gui = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.wav", "*.mp3", "*.png", "*.icns", "*.ico"]
